+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'cookbooks@chef.io'
 license 'Apache 2.0'
 description 'Configures local bootstrapping and installing operating systems via PXE booting.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '2.0.0'
+version '2.1.0'
 depends 'apache2', '>= 1.6'
 depends 'tftp'
 depends 'apt'