+++ recipes/server.rb
@@ -146,7 +146,11 @@ end
 
 # generate local mirror of installers
 include_recipe 'pxe_dust::installers'
-# run dhcpd setup
-include_recipe 'pxe_dust::dhcpd'
+
+if node['pxe_dust']['dhcpd_server']
+  # run dhcpd setup
+  include_recipe 'pxe_dust::dhcpd'
+end
+
 # generate local mirror install.sh and bootstrap templates
 include_recipe 'pxe_dust::bootstrap_template'